@@ -1,7 +1,7 @@
 ﻿using BooksLib;
+
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Design;
-using System;
 
 public class BooksContextFactory : IDesignTimeDbContextFactory<BooksContext>
 {
