@@ -91,9 +91,14 @@ class Runner
     public async Task UpdateRecordsAsync()
     {
         using var context = _menusContextFactory.CreateDbContext();
-        MenuItem menu = await context.MenusItems
+        MenuItem? menu = await context.MenusItems
           .Skip(1)
           .FirstOrDefaultAsync();
+        if (menu is null)
+        {
+            Console.WriteLine("no menu available");
+            return;
+        }
 
         ShowState(context);
         menu.Price += 0.2m;
